requirements.txt
apache-airflow==2.7.3
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
pyarrow==15.0.2
